[Cloud Security] 3P callout displayed in tables (#196335)
(cherry picked from commit 9c2a0418f51bb87f130c3ac7d139bad4d1aa7cc5)
